Move setupNotes to BaseSheetActivity
